move : Search와 CharacterDetail View 나누기, API통신은 APIRemote로 이동
